+++ lib/src/overlays/tools/ruler_tool.dart
@@ -287,6 +287,16 @@ class RulerTool extends ChartOverlay {
   }
 
   @override
+  Rect? selectionBounds(PainterParams params) {
+    final x1 = params.fitTimestamp(startTime) ?? 0;
+    final x2 = params.fitTimestamp(endTime) ?? params.chartWidth;
+    return Rect.fromPoints(
+      Offset(x1, params.fitPrice(startPrice)),
+      Offset(x2, params.fitPrice(endPrice)),
+    );
+  }
+
+  @override
   bool hitTest(Offset position, PainterParams params) {
     if (!interactive) return false;